move main branch to console version; add JUnit dependency; minor code refactor; refactor project file structure.
