--- src/main/java/com/andersentask/bookshop/book/exceptions/NoSuchBookException.java
@@ -1,14 +1,0 @@
-package com.andersentask.bookshop.book.exceptions;
-
-
-public class NoSuchBookException extends RuntimeException {
-    private final static String DEFAULT_MESSAGE = "No such book was found";
-
-    public NoSuchBookException(){
-        super(DEFAULT_MESSAGE);
-    }
-
-    public NoSuchBookException(String message){
-        super(message);
-    }
-}
