--- src/main/java/com/andersentask/bookshop/book/services/interfaces/BookService.java
@@ -1,19 +1,0 @@
-package com.andersentask.bookshop.book.services.interfaces;
-
-import com.andersentask.bookshop.book.dtos.BookDTO;
-import com.andersentask.bookshop.book.entities.Book;
-
-import java.util.List;
-
-public interface BookService {
-
-    Book save(Book book);
-
-    List<BookDTO> getAllBooks();
-
-    List<BookDTO> getBooksSortedByName();
-
-    List<BookDTO> getBooksSortedByPrice();
-
-    List<BookDTO> getBooksSortedByAvailability();
-}
