--- src/main/java/com/andersentask/bookshop/user/domain/dto/UserDto.java
@@ -1,21 +1,0 @@
-package com.andersentask.bookshop.user.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@Builder
-@NoArgsConstructor
-public class UserDto {
-
-    private Long id;
-
-    private String firstName;
-
-    private String lastName;
-
-    private String email;
-}
