--- src/main/java/com/andersentask/bookshop/user/domain/model/Role.java
@@ -1,5 +1,0 @@
-package com.andersentask.bookshop.user.domain.model;
-
-public enum Role {
-    ROLE_USER, ROLE_ADMIN
-}
