+++ src/main/java/com/andersentask/bookshop/user/enums/Role.java
@@ -1,0 +1,5 @@
+package com.andersentask.bookshop.user.enums;
+
+public enum Role {
+    ROLE_USER, ROLE_ADMIN
+}
