+++ src/main/java/com/andersentask/bookshop/user/repository/interfaces/UserCollectionRepository.java
@@ -1,7 +1,7 @@
 package com.andersentask.bookshop.user.repository.interfaces;
 
 import com.andersentask.bookshop.common.AbstractCollectionRepository;
-import com.andersentask.bookshop.user.domain.model.User;
+import com.andersentask.bookshop.user.entities.User;
 
 import java.util.Optional;
