+++ src/main/java/com/andersentask/bookshop/user/service/UserService.java
@@ -1,8 +1,6 @@
 package com.andersentask.bookshop.user.service;
 
-import com.andersentask.bookshop.user.domain.dto.UserDto;
-import com.andersentask.bookshop.user.domain.mapper.UserMapper;
-import com.andersentask.bookshop.user.domain.model.User;
+import com.andersentask.bookshop.user.entities.User;
 import com.andersentask.bookshop.user.repository.UserCollectionRepositoryImpl;
 import lombok.RequiredArgsConstructor;
 
@@ -13,17 +11,17 @@ public class UserService {
 
     public boolean registration(User user) {
         if (userRepository.findByEmailIgnoreCase(user.getEmail()).isPresent()) {
-            System.out.println("User already exist with email:" + user.getEmail());
+            System.out.println("User already exist with email:" + user.getEmail()); // todo: replace with logger
             return false;
         }
         userRepository.save(user);
-        System.out.println("Successfully registered! Your email: " + user.getEmail() + "; password: " + user.getPassword());
+        System.out.println("Successfully registered! Your email: " + user.getEmail() + "; password: " + user.getPassword()); // todo: replace with logger
         return true;
     }
 
-    public UserDto findByEmail(String email) {
-        return UserMapper.entityToDto(userRepository.findByEmailIgnoreCase(email)
-                .orElseThrow(() -> new RuntimeException("User is missing with email:")));
+    public User findByEmail(String email) {
+        return userRepository.findByEmailIgnoreCase(email)
+                .orElseThrow(() -> new RuntimeException("User is missing with email:"));
     }
 
 }
