--- src/test/java/com/andersentask/bookshop/BookshopApplicationTests.java
@@ -1,13 +1,0 @@
-package com.andersentask.bookshop;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BookshopApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
